(admin) move switches CSV import
